clean up collaborator feature (#510)
* clean up collaborator feature

* cleanup  more instances of addCollaborator

* clean up: un-used functions

* cleanup: getReadPoolUpdate

* fix lint errors

* cleanup file: zboxcli_collaborator_test.go

* Remove system test which tests creation of allocation with name (#526)

* Remove system test which tests creation of allocation with name

* Remove tests that verifies update allocation name

* Miner block rewards test (#491)

* Remove no_username (#528)

* remove no_username

* lint fix

* fix(zcncore): used zcncore.MakeSCRestAPICall instead of zboxutil.MakeSCRestAPICall (#514)

* fix(zcncore): used zcncore.MakeSCRestAPICall instead of zboxutil.MakeSCRestAPICall

* fix(upload): improved error message on os.Open

* System test 508 - 0Box Allocation system tests (#519)

* created list and create allocation method for 0box allocation test

* modified allocation struct

* refactor

* remove unused import

* completed the issue

* removed lint issues

* Fix panic in 0box tests (#507)

Add nil check in 0box tests

* Skip tests due to slow goerli

* Worked upon review provided

* created get allocation method

* added one ore test case

* rebasing

* added testcases for get allocation

* added few more testcases

* worked upon review comment

* corrected spelling which was causing lint issues in system tests

* worked upon review and added update allocation

* ready to merge

* ready to merge

---------


* fix lint issues

---------
