+++ tests/cli_tests/flaky___broken_scenarios_test.go
@@ -10,7 +10,6 @@ import (
 	"strconv"
 	"strings"
 	"testing"
-	"time"
 
 	"github.com/0chain/system_test/internal/api/util/test"
 
@@ -329,82 +328,6 @@ func Test___FlakyBrokenScenarios(testSetup *testing.T) {
 
 		require.Greater(t, includedFile_final.Size, includedFile_initial.Size, "included file expected to be updated to bigger size")
 		require.Equal(t, excludedFile_initial.Size, excludedFile_final.Size, "excluded file expected to NOT be updated")
-	})
-
-	// The test is failing due to a possible bug.
-	// When owner downloads the file the cost is deduced from the read pool,
-	// But it seems the collaborators can download the file for free
-	t.RunWithTimeout("Add Collaborator _ file owner must pay for collaborators' reads", 60*time.Second, func(t *test.SystemTest) {
-		collaboratorWalletName := escapedTestName(t) + "_collaborator"
-
-		output, err := registerWalletForName(t, configPath, collaboratorWalletName)
-		require.Nil(t, err, "Unexpected register wallet failure", strings.Join(output, "\n"))
-
-		collaboratorWallet, err := getWalletForName(t, configPath, collaboratorWalletName)
-		require.Nil(t, err, "Error occurred when retrieving curator wallet")
-
-		allocationID := setupAllocation(t, configPath, map[string]interface{}{"size": 4 * MB})
-		createAllocationTestTeardown(t, allocationID)
-
-		localpath := uploadRandomlyGeneratedFile(t, allocationID, "/", 1*MB)
-		remotepath := "/" + filepath.Base(localpath)
-
-		output, err = getDownloadCost(t, configPath, createParams(map[string]interface{}{
-			"allocation": allocationID,
-			"remotepath": remotepath,
-		}), true)
-		require.Nil(t, err, "Could not get download cost", strings.Join(output, "\n"))
-
-		expectedDownloadCostInZCN, err := strconv.ParseFloat(strings.Fields(output[0])[0], 64)
-		require.Nil(t, err, "Cost couldn't be parsed to float", strings.Join(output, "\n"))
-
-		unit := strings.Fields(output[0])[1]
-		expectedDownloadCostInZCN = unitToZCN(expectedDownloadCostInZCN, unit)
-		expectedDownloadCost := ConvertToValue(expectedDownloadCostInZCN)
-
-		output, err = addCollaborator(t, createParams(map[string]interface{}{
-			"allocation": allocationID,
-			"collabid":   collaboratorWallet.ClientID,
-			"remotepath": remotepath,
-		}), true)
-		require.Nil(t, err, "error in adding collaborator", strings.Join(output, "\n"))
-
-		meta := getMetaData(t, map[string]interface{}{
-			"allocation": allocationID,
-			"remotepath": remotepath,
-			"json":       "",
-		})
-		require.Equal(t, 1, len(meta.Collaborators), "Collaborator must be added in file collaborators list")
-		require.Equal(t, collaboratorWallet.ClientID, meta.Collaborators[0].ClientID, "Collaborator must be added in file collaborators list")
-
-		readPoolParams := createParams(map[string]interface{}{
-			"tokens": 0.4,
-		})
-		output, err = readPoolLock(t, configPath, readPoolParams, true)
-		require.Nil(t, err, "Tokens could not be locked", strings.Join(output, "\n"))
-		require.Len(t, output, 1, "Unexpected number of output lines", strings.Join(output, "\n"))
-		require.Equal(t, "locked", output[0])
-
-		readPool := getReadPoolInfo(t)
-		require.Equal(t, ConvertToValue(0.4), readPool.Balance, "Read Pool balance must be equal to locked amount")
-
-		output, err = downloadFileForWallet(t, collaboratorWalletName, configPath, createParams(map[string]interface{}{
-			"allocation": allocationID,
-			"remotepath": remotepath,
-			"localpath":  "tmp/",
-		}), true)
-		require.Nil(t, err, "Error in downloading the file as collaborator", strings.Join(output, "\n"))
-		defer os.Remove("tmp" + remotepath)
-		require.Equal(t, 2, len(output), "Unexpected number of output lines", strings.Join(output, "\n"))
-		expectedOutput := fmt.Sprintf("Status completed callback. Type = application/octet-stream. Name = %s", filepath.Base(localpath))
-		require.Equal(t, expectedOutput, output[1], "Unexpected output", strings.Join(output, "\n"))
-
-		// expected download cost times to the number of blobbers
-		expectedPoolBalance := ConvertToValue(0.4) - expectedDownloadCost
-
-		updatedReadPool, err := getReadPoolUpdate(t, readPool, 5)
-		require.NoError(t, err)
-		require.Equal(t, expectedPoolBalance, updatedReadPool.Balance, "Read Pool balance must be equal to (initial balance-download cost)")
 	})
 
 	// FIXME: WRITEPOOL TOKEN ACCOUNTING
