--- tests/cli_tests/zboxcli_collaborator_test.go
@@ -1,807 +1,0 @@
-package cli_tests
-
-import (
-	"encoding/json"
-	"fmt"
-	"os"
-	"path/filepath"
-	"regexp"
-	"strings"
-	"testing"
-	"time"
-
-	"github.com/0chain/system_test/internal/api/util/test"
-
-	climodel "github.com/0chain/system_test/internal/cli/model"
-	cliutils "github.com/0chain/system_test/internal/cli/util"
-	"github.com/stretchr/testify/require"
-)
-
-func TestCollaborator(testSetup *testing.T) {
-	//todo: some slow operations in here
-	t := test.NewSystemTest(testSetup)
-
-	t.Parallel()
-
-	t.Run("Add Collaborator _ collaborator client id must be added to file collaborators list", func(t *test.SystemTest) {
-		collaboratorWalletName := escapedTestName(t) + "_collaborator"
-
-		output, err := registerWalletForName(t, configPath, collaboratorWalletName)
-		require.Nil(t, err, "Unexpected register wallet failure", strings.Join(output, "\n"))
-
-		collaboratorWallet, err := getWalletForName(t, configPath, collaboratorWalletName)
-		require.Nil(t, err, "Error occurred when retrieving curator wallet")
-
-		allocationID := setupAllocation(t, configPath, map[string]interface{}{"size": 2 * MB})
-		createAllocationTestTeardown(t, allocationID)
-
-		localpath := uploadRandomlyGeneratedFile(t, allocationID, "/", 128*KB)
-		remotepath := "/" + filepath.Base(localpath)
-
-		output, err = addCollaborator(t, createParams(map[string]interface{}{
-			"allocation": allocationID,
-			"collabid":   collaboratorWallet.ClientID,
-			"remotepath": remotepath,
-		}), true)
-		require.Nil(t, err, "error in adding collaborator", strings.Join(output, "\n"))
-		require.Len(t, output, 1, strings.Join(output, "\n"))
-		expectedOutput := fmt.Sprintf("Collaborator %s added successfully for the file %s", collaboratorWallet.ClientID, remotepath)
-		require.Equal(t, expectedOutput, output[0], strings.Join(output, "\n"))
-
-		meta := getMetaData(t, map[string]interface{}{
-			"allocation": allocationID,
-			"remotepath": remotepath,
-			"json":       "",
-		})
-		require.Len(t, meta.Collaborators, 1, "Collaborator must be added in file collaborators list")
-		require.Equal(t, collaboratorWallet.ClientID, meta.Collaborators[0].ClientID, "Collaborator must be added in file collaborators list")
-	})
-
-	t.Run("Add Collaborator _ collaborator can NOT be added to a directory", func(t *test.SystemTest) {
-		collaboratorWalletName := escapedTestName(t) + "_collaborator"
-
-		output, err := registerWalletForName(t, configPath, collaboratorWalletName)
-		require.Nil(t, err, "Unexpected register wallet failure", strings.Join(output, "\n"))
-
-		collaboratorWallet, err := getWalletForName(t, configPath, collaboratorWalletName)
-		require.Nil(t, err, "Error occurred when retrieving curator wallet")
-
-		allocationID := setupAllocation(t, configPath, map[string]interface{}{"size": 2 * MB})
-		createAllocationTestTeardown(t, allocationID)
-
-		dirName := "/NewFolder"
-		output, err = createDir(t, configPath, allocationID, dirName, true)
-		require.Nil(t, err, "Unexpected create dir failure %s", strings.Join(output, "\n"))
-		require.Len(t, output, 1)
-		require.Equal(t, "/NewFolder directory created", output[0])
-
-		meta := getMetaData(t, map[string]interface{}{
-			"allocation": allocationID,
-			"remotepath": dirName,
-			"json":       "",
-		})
-		require.Len(t, meta.Collaborators, 0, "Directory collaborators list expected to be empty")
-
-		// Add collaborator to directory
-		output, err = addCollaborator(t, createParams(map[string]interface{}{
-			"allocation": allocationID,
-			"collabid":   collaboratorWallet.ClientID,
-			"remotepath": dirName,
-		}), false)
-		require.NotNil(t, err, "Unexpected success in adding collaborator", strings.Join(output, "\n"))
-		require.Len(t, output, 1, "Unexpected output length", strings.Join(output, "\n"))
-		expectedOutput := "add_collaborator_failed: Failed to add collaborator on all blobbers."
-		require.Equal(t, expectedOutput, output[0], "Unexpected output in add collaborator", strings.Join(output, "\n"))
-	})
-
-	t.RunWithTimeout("Add Collaborator _ collaborator must be able to read the file", 2*time.Minute, func(t *test.SystemTest) { // todo: too slow
-		collaboratorWalletName := escapedTestName(t) + "_collaborator"
-
-		registerWalletForNameAndLockReadTokens(t, configPath, collaboratorWalletName)
-
-		collaboratorWallet, err := getWalletForName(t, configPath, collaboratorWalletName)
-		require.Nil(t, err, "Error occurred when retrieving curator wallet")
-
-		err = lockReadTokensForWalletName(t, configPath, collaboratorWalletName, 2)
-		require.Nil(t, err)
-
-		allocationID := setupAllocation(t, configPath, map[string]interface{}{"size": 2 * MB})
-		createAllocationTestTeardown(t, allocationID)
-
-		localpath := uploadRandomlyGeneratedFile(t, allocationID, "/", 128*KB)
-		remotepath := "/" + filepath.Base(localpath)
-
-		output, err := addCollaborator(t, createParams(map[string]interface{}{
-			"allocation": allocationID,
-			"collabid":   collaboratorWallet.ClientID,
-			"remotepath": remotepath,
-		}), true)
-		require.Nil(t, err, "error in adding collaborator", strings.Join(output, "\n"))
-
-		meta := getMetaData(t, map[string]interface{}{
-			"allocation": allocationID,
-			"remotepath": remotepath,
-			"json":       "",
-		})
-		require.Len(t, meta.Collaborators, 1, "Collaborator must be added in file collaborators list")
-		require.Equal(t, collaboratorWallet.ClientID, meta.Collaborators[0].ClientID, "Collaborator must be added in file collaborators list")
-
-		readPoolParams := createParams(map[string]interface{}{
-			"tokens": 0.4,
-		})
-		output, err = readPoolLock(t, configPath, readPoolParams, true)
-		require.Nil(t, err, "Tokens could not be locked", strings.Join(output, "\n"))
-		require.Len(t, output, 1, "Unexpected number of output lines", strings.Join(output, "\n"))
-		require.Equal(t, "locked", output[0])
-
-		readPool := getReadPoolInfo(t)
-		require.Equal(t, ConvertToValue(0.4), readPool.Balance, "Read Pool balance must be equal to locked amount")
-
-		output, err = downloadFileForWallet(t, collaboratorWalletName, configPath, createParams(map[string]interface{}{
-			"allocation": allocationID,
-			"remotepath": remotepath,
-			"localpath":  "tmp/",
-		}), true)
-
-		require.Nil(t, err, "Error in downloading the file as collaborator", strings.Join(output, "\n"))
-		defer os.Remove("tmp" + remotepath)
-		require.Len(t, output, 2, "Unexpected number of output lines", strings.Join(output, "\n"))
-		require.Contains(t, output[1], StatusCompletedCB)
-		require.Contains(t, output[1], filepath.Base(localpath))
-	})
-
-	t.Run("Add Collaborator _ collaborator must not be able to share the file", func(t *test.SystemTest) {
-		collaboratorWalletName := escapedTestName(t) + "_collaborator"
-
-		output, err := registerWalletForName(t, configPath, collaboratorWalletName)
-		require.Nil(t, err, "Unexpected register wallet failure", strings.Join(output, "\n"))
-
-		collaboratorWallet, err := getWalletForName(t, configPath, collaboratorWalletName)
-		require.Nil(t, err, "Error occurred when retrieving curator wallet")
-
-		allocationID := setupAllocation(t, configPath, map[string]interface{}{"size": 2 * MB})
-		createAllocationTestTeardown(t, allocationID)
-
-		localpath := uploadRandomlyGeneratedFile(t, allocationID, "/", 128*KB)
-		remotepath := "/" + filepath.Base(localpath)
-
-		output, err = addCollaborator(t, createParams(map[string]interface{}{
-			"allocation": allocationID,
-			"collabid":   collaboratorWallet.ClientID,
-			"remotepath": remotepath,
-		}), true)
-		require.Nil(t, err, "error in adding collaborator", strings.Join(output, "\n"))
-
-		meta := getMetaData(t, map[string]interface{}{
-			"allocation": allocationID,
-			"remotepath": remotepath,
-			"json":       "",
-		})
-		require.Len(t, meta.Collaborators, 1, "Collaborator must be added in file collaborators list")
-		require.Equal(t, collaboratorWallet.ClientID, meta.Collaborators[0].ClientID, "Collaborator must be added in file collaborators list")
-
-		output, err = shareFileWithWallet(t, collaboratorWalletName, configPath, map[string]interface{}{
-			"allocation": allocationID,
-			"remotepath": remotepath,
-		})
-		require.NotNil(t, err, "expected error when collaborator shares a file")
-		require.Len(t, output, 1)
-		require.Regexp(t, regexp.MustCompile("consensus not reached"), output[0], "Unexpected output", strings.Join(output, "\n"))
-	})
-
-	t.Run("Remove Collaborator _ collaborator client id must be removed from file collaborators list", func(t *test.SystemTest) {
-		collaboratorWalletName := escapedTestName(t) + "_collaborator"
-
-		output, err := registerWalletForName(t, configPath, collaboratorWalletName)
-		require.Nil(t, err, "Unexpected register wallet failure", strings.Join(output, "\n"))
-
-		collaboratorWallet, err := getWalletForName(t, configPath, collaboratorWalletName)
-		require.Nil(t, err, "Error occurred when retrieving curator wallet")
-
-		allocationID := setupAllocation(t, configPath, map[string]interface{}{"size": 2 * MB})
-		createAllocationTestTeardown(t, allocationID)
-
-		localpath := uploadRandomlyGeneratedFile(t, allocationID, "/", 128*KB)
-		remotepath := "/" + filepath.Base(localpath)
-
-		output, err = addCollaborator(t, createParams(map[string]interface{}{
-			"allocation": allocationID,
-			"collabid":   collaboratorWallet.ClientID,
-			"remotepath": remotepath,
-		}), true)
-		require.Nil(t, err, "error in adding collaborator", strings.Join(output, "\n"))
-		require.Len(t, output, 1, strings.Join(output, "\n"))
-		expectedOutput := fmt.Sprintf("Collaborator %s added successfully for the file %s", collaboratorWallet.ClientID, remotepath)
-		require.Equal(t, expectedOutput, output[0], strings.Join(output, "\n"))
-
-		meta := getMetaData(t, map[string]interface{}{
-			"allocation": allocationID,
-			"remotepath": remotepath,
-			"json":       "",
-		})
-		require.Len(t, meta.Collaborators, 1, "Collaborator must be added in file collaborators list")
-		require.Equal(t, collaboratorWallet.ClientID, meta.Collaborators[0].ClientID, "Collaborator must be added in file collaborators list")
-
-		output, err = removeCollaborator(t, createParams(map[string]interface{}{
-			"allocation": allocationID,
-			"collabid":   collaboratorWallet.ClientID,
-			"remotepath": remotepath,
-		}), true)
-		require.Nil(t, err, "error in deleting collaborator", strings.Join(output, "\n"))
-		require.Len(t, output, 1, strings.Join(output, "\n"))
-		expectedOutput = fmt.Sprintf("Collaborator %s removed successfully for the file %s", collaboratorWallet.ClientID, remotepath)
-		require.Equal(t, expectedOutput, output[0], strings.Join(output, "\n"))
-
-		meta = getMetaData(t, map[string]interface{}{
-			"allocation": allocationID,
-			"remotepath": remotepath,
-			"json":       "",
-		})
-		require.Len(t, meta.Collaborators, 0, "Collaborator must be removed from file collaborators list")
-	})
-
-	t.RunWithTimeout("Remove Collaborator _ file shouldn't be accessible by collaborator anymore", 60*time.Second, func(t *test.SystemTest) {
-		collaboratorWalletName := escapedTestName(t) + "_collaborator"
-
-		output, err := registerWalletForName(t, configPath, collaboratorWalletName)
-		require.Nil(t, err, "Unexpected register wallet failure", strings.Join(output, "\n"))
-
-		collaboratorWallet, err := getWalletForName(t, configPath, collaboratorWalletName)
-		require.Nil(t, err, "Error occurred when retrieving curator wallet")
-
-		allocationID := setupAllocation(t, configPath, map[string]interface{}{"size": 2 * MB})
-		createAllocationTestTeardown(t, allocationID)
-
-		localpath := uploadRandomlyGeneratedFile(t, allocationID, "/", 128*KB)
-		remotepath := "/" + filepath.Base(localpath)
-
-		output, err = addCollaborator(t, createParams(map[string]interface{}{
-			"allocation": allocationID,
-			"collabid":   collaboratorWallet.ClientID,
-			"remotepath": remotepath,
-		}), true)
-		require.Nil(t, err, "error in adding collaborator", strings.Join(output, "\n"))
-
-		meta := getMetaData(t, map[string]interface{}{
-			"allocation": allocationID,
-			"remotepath": remotepath,
-			"json":       "",
-		})
-		require.Len(t, meta.Collaborators, 1, "Collaborator must be added in file collaborators list")
-		require.Equal(t, collaboratorWallet.ClientID, meta.Collaborators[0].ClientID, "Collaborator must be added in file collaborators list")
-
-		// Lock tokens in read pool
-		readPoolParams := createParams(map[string]interface{}{
-			"tokens": 0.4,
-		})
-		output, err = readPoolLock(t, configPath, readPoolParams, true)
-		require.Nil(t, err, "Tokens could not be locked", strings.Join(output, "\n"))
-		require.Len(t, output, 1, "Unexpected number of output lines", strings.Join(output, "\n"))
-		require.Equal(t, "locked", output[0])
-
-		output, err = removeCollaborator(t, createParams(map[string]interface{}{
-			"allocation": allocationID,
-			"collabid":   collaboratorWallet.ClientID,
-			"remotepath": remotepath,
-		}), true)
-		require.Nil(t, err, "error in deleting collaborator", strings.Join(output, "\n"))
-
-		meta = getMetaData(t, map[string]interface{}{
-			"allocation": allocationID,
-			"remotepath": remotepath,
-			"json":       "",
-		})
-		require.Len(t, meta.Collaborators, 0, "Collaborator must be removed from file collaborators list")
-
-		output, err = downloadFileForWallet(t, collaboratorWalletName, configPath, createParams(map[string]interface{}{
-			"allocation": allocationID,
-			"remotepath": remotepath,
-			"localpath":  "tmp/",
-		}), false)
-		require.NotNil(t, err, "The command must fail since the wallet is not collaborator anymore", strings.Join(output, "\n"))
-		require.Len(t, output, 1, "Unexpected number of output lines", strings.Join(output, "\n"))
-		require.Contains(t, output[0], "consensus_not_met")
-		require.Contains(t, output[0], "file meta data")
-	})
-
-	t.Run("Add Collaborator to a file owned by somebody else must fail", func(t *test.SystemTest) {
-		ownerWalletName := escapedTestName(t) + "_owner"
-		anotherWalletName := escapedTestName(t) + "_another"
-
-		allocationID := setupAllocationWithWallet(t, ownerWalletName, configPath, map[string]interface{}{"size": 2 * MB})
-		createAllocationTestTeardown(t, allocationID)
-
-		output, err := registerWalletForName(t, configPath, anotherWalletName)
-		require.Nil(t, err, "Unexpected register wallet failure", strings.Join(output, "\n"))
-
-		output, err = executeFaucetWithTokensForWallet(t, anotherWalletName, configPath, 1.0)
-		require.Nil(t, err, "faucet execution failed", strings.Join(output, "\n"))
-
-		anotherWallet, err := getWalletForName(t, configPath, anotherWalletName)
-		require.Nil(t, err, "Error occurred when retrieving curator wallet")
-
-		localpath := uploadRandomlyGeneratedFileWithWallet(t, ownerWalletName, allocationID, "/", 128*KB)
-		remotepath := "/" + filepath.Base(localpath)
-
-		output, err = addCollaboratorWithWallet(t, anotherWalletName, createParams(map[string]interface{}{
-			"allocation": allocationID,
-			"collabid":   anotherWallet.ClientID,
-			"remotepath": remotepath,
-		}), false)
-		require.NotNil(t, err, "Add collaborator must fail since the wallet is not the file owner", strings.Join(output, "\n"))
-		require.Len(t, output, 1, "Unexpected number of output lines", strings.Join(output, "\n"))
-		require.Equal(t, "add_collaborator_failed: Failed to add collaborator on all blobbers.", output[0], "Unexpected output", strings.Join(output, "\n"))
-	})
-
-	t.RunWithTimeout("Remove Collaborator from a file owned by somebody else must fail", 60*time.Second, func(t *test.SystemTest) {
-		ownerWalletName := escapedTestName(t) + "_owner"
-		anotherWalletName := escapedTestName(t) + "_another"
-
-		allocationID := setupAllocationWithWallet(t, ownerWalletName, configPath, map[string]interface{}{"size": 2 * MB})
-		createAllocationTestTeardown(t, allocationID)
-
-		output, err := registerWalletForName(t, configPath, anotherWalletName)
-		require.Nil(t, err, "Unexpected register wallet failure", strings.Join(output, "\n"))
-
-		output, err = executeFaucetWithTokensForWallet(t, anotherWalletName, configPath, 1.0)
-		require.Nil(t, err, "faucet execution failed", strings.Join(output, "\n"))
-
-		localpath := uploadRandomlyGeneratedFileWithWallet(t, ownerWalletName, allocationID, "/", 128*KB)
-		remotepath := "/" + filepath.Base(localpath)
-
-		thirdPersonWalletAddress := "someone_wallet_address"
-
-		output, err = addCollaboratorWithWallet(t, ownerWalletName, createParams(map[string]interface{}{
-			"allocation": allocationID,
-			"collabid":   thirdPersonWalletAddress,
-			"remotepath": remotepath,
-		}), true)
-		require.Nil(t, err, "error in adding collaborator", strings.Join(output, "\n"))
-		require.Len(t, output, 1, strings.Join(output, "\n"))
-		expectedOutput := fmt.Sprintf("Collaborator %s added successfully for the file %s", thirdPersonWalletAddress, remotepath)
-		require.Equal(t, expectedOutput, output[0], strings.Join(output, "\n"))
-
-		meta := getMetaDataWithWallet(t, ownerWalletName, map[string]interface{}{
-			"allocation": allocationID,
-			"remotepath": remotepath,
-			"json":       "",
-		})
-		require.Len(t, meta.Collaborators, 1, "Collaborator must be added in file collaborators list")
-		require.Equal(t, thirdPersonWalletAddress, meta.Collaborators[0].ClientID, "Collaborator must be added in file collaborators list")
-
-		// Now we test if another wallet can remove from collaborators' list
-		output, err = removeCollaboratorWithWallet(t, anotherWalletName, createParams(map[string]interface{}{
-			"allocation": allocationID,
-			"collabid":   thirdPersonWalletAddress,
-			"remotepath": remotepath,
-		}), false)
-		require.NotNil(t, err, "Remove collaborator must fail since the wallet is not the file owner", strings.Join(output, "\n"))
-		require.Len(t, output, 1, "Unexpected number of output lines", strings.Join(output, "\n"))
-		require.Equal(t, "remove_collaborator_failed: Failed to remove collaborator on all blobbers.", output[0], "Unexpected output", strings.Join(output, "\n"))
-	})
-
-	t.Run("Add Collaborator _ Collaborator should NOT be able to add another collaborator", func(t *test.SystemTest) {
-		collaboratorWalletName := escapedTestName(t) + "_collaborator"
-
-		allocationID := setupAllocation(t, configPath, map[string]interface{}{"size": 2 * MB})
-		createAllocationTestTeardown(t, allocationID)
-
-		output, err := registerWalletForName(t, configPath, collaboratorWalletName)
-		require.Nil(t, err, "Unexpected register wallet failure", strings.Join(output, "\n"))
-
-		collaboratorWallet, err := getWalletForName(t, configPath, collaboratorWalletName)
-		require.Nil(t, err, "error in getting wallet", err, strings.Join(output, "\n"))
-
-		output, err = executeFaucetWithTokens(t, configPath, 1.0)
-		require.Nil(t, err, "faucet execution failed", strings.Join(output, "\n"))
-
-		localpath := uploadRandomlyGeneratedFile(t, allocationID, "/", 128*KB)
-		remotepath := "/" + filepath.Base(localpath)
-
-		thirdPersonWalletAddress := "someone_wallet_address"
-
-		output, err = addCollaborator(t, createParams(map[string]interface{}{
-			"allocation": allocationID,
-			"collabid":   collaboratorWallet.ClientID,
-			"remotepath": remotepath,
-		}), true)
-		require.Nil(t, err, "error in adding collaborator", strings.Join(output, "\n"))
-		require.Len(t, output, 1, strings.Join(output, "\n"))
-		expectedOutput := fmt.Sprintf("Collaborator %s added successfully for the file %s", collaboratorWallet.ClientID, remotepath)
-		require.Equal(t, expectedOutput, output[0], strings.Join(output, "\n"))
-
-		meta := getMetaData(t, map[string]interface{}{
-			"allocation": allocationID,
-			"remotepath": remotepath,
-			"json":       "",
-		})
-		require.Len(t, meta.Collaborators, 1, "Collaborator must be added in file collaborators list")
-		require.Equal(t, collaboratorWallet.ClientID, meta.Collaborators[0].ClientID, "Collaborator must be added in file collaborators list")
-
-		// Now we test if collaborator can add another collaborator to filr
-		output, err = addCollaboratorWithWallet(t, collaboratorWalletName, createParams(map[string]interface{}{
-			"allocation": allocationID,
-			"collabid":   thirdPersonWalletAddress,
-			"remotepath": remotepath,
-		}), false)
-		require.NotNil(t, err, "Add collaborator must fail since the collaborator is not the file owner", strings.Join(output, "\n"))
-		require.Len(t, output, 1, "Unexpected number of output lines", strings.Join(output, "\n"))
-		require.Equal(t, "add_collaborator_failed: Failed to add collaborator on all blobbers.", output[0], "Unexpected output", strings.Join(output, "\n"))
-	})
-
-	t.RunWithTimeout("Add Collaborator _ collaborator should NOT be able to rename the file", 90*time.Second, func(t *test.SystemTest) {
-		collaboratorWalletName := escapedTestName(t) + "_collaborator"
-
-		output, err := registerWalletForName(t, configPath, collaboratorWalletName)
-		require.Nil(t, err, "Unexpected register wallet failure", strings.Join(output, "\n"))
-
-		collaboratorWallet, err := getWalletForName(t, configPath, collaboratorWalletName)
-		require.Nil(t, err, "Error occurred when retrieving curator wallet")
-
-		allocationID := setupAllocation(t, configPath, map[string]interface{}{"size": 2 * MB})
-		createAllocationTestTeardown(t, allocationID)
-
-		localpath := uploadRandomlyGeneratedFile(t, allocationID, "/tmp", 128*KB)
-		remotepath := "/tmp/" + filepath.Base(localpath)
-
-		output, err = addCollaborator(t, createParams(map[string]interface{}{
-			"allocation": allocationID,
-			"collabid":   collaboratorWallet.ClientID,
-			"remotepath": remotepath,
-		}), true)
-		require.Nil(t, err, "error in adding collaborator", strings.Join(output, "\n"))
-
-		meta := getMetaData(t, map[string]interface{}{
-			"allocation": allocationID,
-			"remotepath": remotepath,
-			"json":       "",
-		})
-		require.Len(t, meta.Collaborators, 1, "Collaborator must be added in file collaborators list")
-		require.Equal(t, collaboratorWallet.ClientID, meta.Collaborators[0].ClientID, "Collaborator must be added in file collaborators list")
-
-		output, err = renameFileWithWallet(t, configPath, collaboratorWalletName, map[string]interface{}{
-			"allocation": allocationID,
-			"remotepath": remotepath,
-			"destname":   "new_name.txt",
-		})
-		require.NotNil(t, err, "Unexpected success in renaming the file as collaborator", strings.Join(output, "\n"))
-		require.Len(t, output, 1, "Unexpected number of output lines", strings.Join(output, "\n"))
-		require.Contains(t, output[0], "consensus_not_met")
-	})
-
-	t.Run("Add Collaborator _ collaborator should NOT be able to delete the file", func(t *test.SystemTest) {
-		collaboratorWalletName := escapedTestName(t) + "_collaborator"
-
-		output, err := registerWalletForName(t, configPath, collaboratorWalletName)
-		require.Nil(t, err, "Unexpected register wallet failure", strings.Join(output, "\n"))
-
-		collaboratorWallet, err := getWalletForName(t, configPath, collaboratorWalletName)
-		require.Nil(t, err, "Error occurred when retrieving curator wallet")
-
-		allocationID := setupAllocation(t, configPath, map[string]interface{}{"size": 2 * MB})
-		createAllocationTestTeardown(t, allocationID)
-
-		localpath := uploadRandomlyGeneratedFile(t, allocationID, "/tmp", 128*KB)
-		remotepath := "/tmp/" + filepath.Base(localpath)
-
-		output, err = addCollaborator(t, createParams(map[string]interface{}{
-			"allocation": allocationID,
-			"collabid":   collaboratorWallet.ClientID,
-			"remotepath": remotepath,
-		}), true)
-		require.Nil(t, err, "error in adding collaborator", strings.Join(output, "\n"))
-
-		meta := getMetaData(t, map[string]interface{}{
-			"allocation": allocationID,
-			"remotepath": remotepath,
-			"json":       "",
-		})
-		require.Len(t, meta.Collaborators, 1, "Collaborator must be added in file collaborators list")
-		require.Equal(t, collaboratorWallet.ClientID, meta.Collaborators[0].ClientID, "Collaborator must be added in file collaborators list")
-
-		output, err = deleteFile(t, collaboratorWalletName, createParams(map[string]interface{}{
-			"allocation": allocationID,
-			"remotepath": remotepath,
-		}), false)
-		require.NotNil(t, err, "Unexpected success in deleting the file as collaborator", strings.Join(output, "\n"))
-		require.Len(t, output, 1, "Unexpected number of output lines", strings.Join(output, "\n"))
-		require.Contains(t, output[0], "consensus_not_met")
-	})
-
-	t.Run("Add Collaborator _ collaborator should NOT be able to move the file", func(t *test.SystemTest) {
-		collaboratorWalletName := escapedTestName(t) + "_collaborator"
-
-		output, err := registerWalletForName(t, configPath, collaboratorWalletName)
-		require.Nil(t, err, "Unexpected register wallet failure", strings.Join(output, "\n"))
-
-		collaboratorWallet, err := getWalletForName(t, configPath, collaboratorWalletName)
-		require.Nil(t, err, "Error occurred when retrieving curator wallet")
-
-		allocationID := setupAllocation(t, configPath, map[string]interface{}{"size": 2 * MB})
-		createAllocationTestTeardown(t, allocationID)
-
-		localpath := uploadRandomlyGeneratedFile(t, allocationID, "/tmp", 128*KB)
-		remotepath := "/tmp/" + filepath.Base(localpath)
-
-		output, err = addCollaborator(t, createParams(map[string]interface{}{
-			"allocation": allocationID,
-			"collabid":   collaboratorWallet.ClientID,
-			"remotepath": remotepath,
-		}), true)
-		require.Nil(t, err, "error in adding collaborator", strings.Join(output, "\n"))
-
-		meta := getMetaData(t, map[string]interface{}{
-			"allocation": allocationID,
-			"remotepath": remotepath,
-			"json":       "",
-		})
-		require.Len(t, meta.Collaborators, 1, "Collaborator must be added in file collaborators list")
-		require.Equal(t, collaboratorWallet.ClientID, meta.Collaborators[0].ClientID, "Collaborator must be added in file collaborators list")
-
-		output, err = moveFileWithWallet(t, collaboratorWalletName, configPath, map[string]interface{}{
-			"allocation": allocationID,
-			"remotepath": remotepath,
-			"destpath":   "/",
-		}, true)
-		require.Nil(t, err, "Error in moving the file as collaborator", strings.Join(output, "\n"))
-		require.Len(t, output, 1, "Unexpected number of output lines", strings.Join(output, "\n"))
-		require.Contains(t, output[0], "consensus_not_met")
-	})
-
-	t.Run("Add Collaborator _ collaborator should be able to update the file", func(t *test.SystemTest) {
-		collaboratorWalletName := escapedTestName(t) + "_collaborator"
-
-		output, err := registerWalletForName(t, configPath, collaboratorWalletName)
-		require.Nil(t, err, "Unexpected register wallet failure", strings.Join(output, "\n"))
-
-		collaboratorWallet, err := getWalletForName(t, configPath, collaboratorWalletName)
-		require.Nil(t, err, "Error occurred when retrieving curator wallet")
-
-		allocationID := setupAllocation(t, configPath, map[string]interface{}{"size": 2 * MB})
-		createAllocationTestTeardown(t, allocationID)
-
-		localpath := uploadRandomlyGeneratedFile(t, allocationID, "/tmp", 128*KB)
-		remotepath := "/tmp/" + filepath.Base(localpath)
-
-		output, err = addCollaborator(t, createParams(map[string]interface{}{
-			"allocation": allocationID,
-			"collabid":   collaboratorWallet.ClientID,
-			"remotepath": remotepath,
-		}), true)
-		require.Nil(t, err, "error in adding collaborator", strings.Join(output, "\n"))
-
-		meta := getMetaData(t, map[string]interface{}{
-			"allocation": allocationID,
-			"remotepath": remotepath,
-			"json":       "",
-		})
-		require.Len(t, meta.Collaborators, 1, "Collaborator must be added in file collaborators list")
-		require.Equal(t, collaboratorWallet.ClientID, meta.Collaborators[0].ClientID, "Collaborator must be added in file collaborators list")
-
-		updatedLocalPath := generateRandomTestFileName(t)
-		err = createFileWithSize(updatedLocalPath, 128*KB)
-		require.Nil(t, err)
-
-		output, err = updateFileWithWallet(t, collaboratorWalletName, configPath, map[string]interface{}{
-			"allocation": allocationID,
-			"remotepath": remotepath,
-			"localpath":  updatedLocalPath,
-		}, true)
-
-		defer os.Remove(updatedLocalPath)
-		require.Nil(t, err, "failed in updating the file as collaborator", strings.Join(output, "\n"))
-		require.Len(t, output, 2, "Unexpected number of output lines", strings.Join(output, "\n"))
-		require.Contains(t, output[1], StatusCompletedCB)
-		require.Contains(t, output[1], filepath.Base(localpath))
-	})
-
-	t.Run("Add Collaborator _ collaborator should NOT be able to copy the file", func(t *test.SystemTest) {
-		collaboratorWalletName := escapedTestName(t) + "_collaborator"
-
-		output, err := registerWalletForName(t, configPath, collaboratorWalletName)
-		require.Nil(t, err, "Unexpected register wallet failure", strings.Join(output, "\n"))
-
-		collaboratorWallet, err := getWalletForName(t, configPath, collaboratorWalletName)
-		require.Nil(t, err, "Error occurred when retrieving curator wallet")
-
-		allocationID := setupAllocation(t, configPath, map[string]interface{}{"size": 2 * MB})
-		createAllocationTestTeardown(t, allocationID)
-
-		localpath := uploadRandomlyGeneratedFile(t, allocationID, "/tmp", 128*KB)
-		remotepath := "/tmp/" + filepath.Base(localpath)
-
-		output, err = addCollaborator(t, createParams(map[string]interface{}{
-			"allocation": allocationID,
-			"collabid":   collaboratorWallet.ClientID,
-			"remotepath": remotepath,
-		}), true)
-		require.Nil(t, err, "error in adding collaborator", strings.Join(output, "\n"))
-
-		meta := getMetaData(t, map[string]interface{}{
-			"allocation": allocationID,
-			"remotepath": remotepath,
-			"json":       "",
-		})
-		require.Len(t, meta.Collaborators, 1, "Collaborator must be added in file collaborators list")
-		require.Equal(t, collaboratorWallet.ClientID, meta.Collaborators[0].ClientID, "Collaborator must be added in file collaborators list")
-
-		output, err = copyFileForWallet(t, configPath, collaboratorWalletName, map[string]interface{}{
-			"allocation": allocationID,
-			"remotepath": remotepath,
-			"destpath":   "/",
-		}, true)
-		require.Nil(t, err, "Unexpected success in copying the file as collaborator", strings.Join(output, "\n"))
-		require.Len(t, output, 1, "Unexpected number of output lines", strings.Join(output, "\n"))
-		require.Contains(t, output[0], "consensus_not_met")
-	})
-
-	t.RunWithTimeout("Add Collaborator _ collaborator should NOT be able to download encrypted file", 2*time.Minute, func(t *test.SystemTest) { // todo: too slow
-		collaboratorWalletName := escapedTestName(t) + "_collaborator"
-
-		registerWalletForNameAndLockReadTokens(t, configPath, collaboratorWalletName)
-
-		collaboratorWallet, err := getWalletForName(t, configPath, collaboratorWalletName)
-		require.Nil(t, err, "Error occurred when retrieving curator wallet")
-
-		err = lockReadTokensForWalletName(t, configPath, collaboratorWalletName, 2)
-		require.Nil(t, err)
-
-		allocationID := setupAllocation(t, configPath, map[string]interface{}{"size": 2 * MB})
-		createAllocationTestTeardown(t, allocationID)
-
-		localpath := generateRandomTestFileName(t)
-		err = createFileWithSize(localpath, 128*KB)
-		require.Nil(t, err)
-		defer os.Remove(localpath)
-
-		output, err := uploadFile(t, configPath, map[string]interface{}{
-			"allocation": allocationID,
-			"remotepath": "/" + filepath.Base(localpath),
-			"localpath":  localpath,
-			"encrypt":    "",
-		}, true)
-		require.Nil(t, err, strings.Join(output, "\n"))
-		require.Len(t, output, 2)
-		require.Contains(t, output[1], StatusCompletedCB)
-		require.Contains(t, output[1], filepath.Base(localpath))
-
-		remotepath := "/" + filepath.Base(localpath)
-
-		output, err = addCollaborator(t, createParams(map[string]interface{}{
-			"allocation": allocationID,
-			"collabid":   collaboratorWallet.ClientID,
-			"remotepath": remotepath,
-		}), true)
-		require.Nil(t, err, "error in adding collaborator", strings.Join(output, "\n"))
-
-		meta := getMetaData(t, map[string]interface{}{
-			"allocation": allocationID,
-			"remotepath": remotepath,
-			"json":       "",
-		})
-		require.Len(t, meta.Collaborators, 1, "Collaborator must be added in file collaborators list")
-		require.Equal(t, collaboratorWallet.ClientID, meta.Collaborators[0].ClientID, "Collaborator must be added in file collaborators list")
-
-		lockedTokens := 0.4
-		readPoolParams := createParams(map[string]interface{}{
-			"tokens": lockedTokens,
-		})
-		output, err = readPoolLock(t, configPath, readPoolParams, true)
-		require.Nil(t, err, "Tokens could not be locked", strings.Join(output, "\n"))
-		require.Len(t, output, 1, "Unexpected number of output lines", strings.Join(output, "\n"))
-		require.Equal(t, "locked", output[0])
-
-		readPool := getReadPoolInfo(t)
-		require.Equal(t, ConvertToValue(lockedTokens), readPool.Balance, "Read Pool balance must be equal to locked amount")
-
-		output, err = downloadFileForWallet(t, collaboratorWalletName, configPath, createParams(map[string]interface{}{
-			"allocation": allocationID,
-			"remotepath": remotepath,
-			"localpath":  "tmp/",
-		}), true)
-		require.NotNil(t, err, "Unexpected success in downloading the file as collaborator", strings.Join(output, "\n"))
-		require.Len(t, output, 3, "Unexpected number of output lines", strings.Join(output, "\n"))
-		aggregatedOutput := strings.Join(output, " ")
-		require.Contains(t, aggregatedOutput, "decryption_error")
-	})
-}
-
-func getReadPoolUpdate(t *test.SystemTest, erp climodel.ReadPoolInfo, retry int) (climodel.ReadPoolInfo, error) {
-	if retry == 0 {
-		retry = 1
-	}
-	// Wait for read markers to be redeemed
-	for i := 0; i < retry; i++ {
-		readPool := getReadPoolInfo(t)
-		if readPool.Balance == erp.Balance {
-			continue
-		}
-
-		cliutils.Wait(t, time.Second*30)
-		return getReadPoolInfo(t), nil
-	}
-
-	return erp, fmt.Errorf("no update found in readpool")
-}
-
-func getReadPoolInfo(t *test.SystemTest) climodel.ReadPoolInfo {
-	output, err := readPoolInfo(t, configPath)
-	require.Nil(t, err, "Error fetching read pool", strings.Join(output, "\n"))
-	require.Len(t, output, 1)
-
-	var readPool climodel.ReadPoolInfo
-	err = json.Unmarshal([]byte(output[0]), &readPool)
-	require.Nil(t, err, "Error unmarshalling read pool %s", strings.Join(output, "\n"))
-	return readPool
-}
-
-func getMetaData(t *test.SystemTest, params map[string]interface{}) *climodel.FileMetaResult {
-	return getMetaDataWithWallet(t, escapedTestName(t), params)
-}
-
-func getMetaDataWithWallet(t *test.SystemTest, walletName string, params map[string]interface{}) *climodel.FileMetaResult {
-	output, err := getFileMetaWithWallet(t, walletName, configPath, createParams(params), true)
-	require.Nil(t, err, "Error in getting file meta data", strings.Join(output, "\n"))
-	require.Len(t, output, 1, "Error in getting file meta data - Unexpected number of output lines", strings.Join(output, "\n"))
-
-	var meta climodel.FileMetaResult
-	err = json.NewDecoder(strings.NewReader(output[0])).Decode(&meta)
-	require.Nil(t, err, "Failed to unmarshal the json result into FileMetaResult", strings.Join(output, "\n"))
-	return &meta
-}
-
-func addCollaborator(t *test.SystemTest, params string, retry bool) ([]string, error) {
-	return addCollaboratorWithWallet(t, escapedTestName(t), params, retry)
-}
-
-func addCollaboratorWithWallet(t *test.SystemTest, walletName, params string, retry bool) ([]string, error) {
-	t.Logf("Adding collaborator...")
-	cmd := fmt.Sprintf(
-		"./zbox add-collab %s --silent --wallet %s "+
-			"--configDir ./config --config %s",
-		params,
-		walletName+"_wallet.json",
-		configPath,
-	)
-	if retry {
-		return cliutils.RunCommand(t, cmd, 3, time.Second*20)
-	} else {
-		return cliutils.RunCommandWithoutRetry(cmd)
-	}
-}
-
-func removeCollaborator(t *test.SystemTest, params string, retry bool) ([]string, error) {
-	return removeCollaboratorWithWallet(t, escapedTestName(t), params, retry)
-}
-
-func removeCollaboratorWithWallet(t *test.SystemTest, walletName, params string, retry bool) ([]string, error) {
-	t.Logf("Removing collaborator...")
-	cmd := fmt.Sprintf(
-		"./zbox delete-collab %s --silent --wallet %s "+
-			"--configDir ./config --config %s",
-		params,
-		walletName+"_wallet.json",
-		configPath,
-	)
-	if retry {
-		return cliutils.RunCommand(t, cmd, 3, time.Second*20)
-	} else {
-		return cliutils.RunCommandWithoutRetry(cmd)
-	}
-}
-
-func deleteFile(t *test.SystemTest, walletName, params string, retry bool) ([]string, error) {
-	t.Logf("Deleting file...")
-	cmd := fmt.Sprintf(
-		"./zbox delete %s --silent --wallet %s "+
-			"--configDir ./config --config %s",
-		params,
-		walletName+"_wallet.json",
-		configPath,
-	)
-	if retry {
-		return cliutils.RunCommand(t, cmd, 3, time.Second*20)
-	} else {
-		return cliutils.RunCommandWithoutRetry(cmd)
-	}
-}
