+++ tests/cli_tests/zwalletcli_register_wallet_test.go
@@ -126,19 +126,6 @@ func registerWalletWithTokens(t *test.SystemTest, cliConfigFilename, name string
 	require.NoErrorf(t, err, "get tokens for wallet %s", name)
 }
 
-func lockReadTokensForWalletName(t *test.SystemTest, cliConfigFilename, wallet string, tokens float64) error {
-	_, err := executeFaucetWithTokensForWallet(t, wallet, cliConfigFilename, tokens)
-	if err != nil {
-		return err
-	}
-
-	readPoolParams := createParams(map[string]interface{}{
-		"tokens": tokens / 2,
-	})
-	_, err = readPoolLockWithWallet(t, wallet, cliConfigFilename, readPoolParams, true)
-	return err
-}
-
 func getBalance(t *test.SystemTest, cliConfigFilename string) ([]string, error) {
 	cliutils.Wait(t, 5*time.Second)
 	return getBalanceForWallet(t, cliConfigFilename, escapedTestName(t))
